+++ pom.xml
@@ -98,6 +98,12 @@
             <artifactId>jboss-logging</artifactId>
             <version>${jboss.logging.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.jboss.maven.plugins</groupId>
+            <artifactId>maven-qstools-plugin</artifactId>
+            <version>1.2.0.CR5</version>
+            <type>maven-plugin</type>
+        </dependency>
     </dependencies>
 
     <build>
