+++ src/main/java/org/jboss/jdf/stacks/model/ArchetypeVersion.java
@@ -14,113 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.jboss.jdf.stacks.model;
 
 import java.util.Properties;
 
-public class ArchetypeVersion {
+public interface ArchetypeVersion {
 
-    private String id;
+    public String getId();
 
-    private Archetype archetype;
+    public Archetype getArchetype();
 
-    private String version;
+    public String getVersion();
 
-    private String repositoryURL;
+    public String getRepositoryURL();
 
-    private Properties labels = new Properties();
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Archetype getArchetype() {
-        return archetype;
-    }
-
-    public void setArchetype(Archetype archetype) {
-        this.archetype = archetype;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getRepositoryURL() {
-        if (this.repositoryURL == null){
-            return getArchetype().getRepositoryURL();
-        }
-        return repositoryURL;
-    }
-
-    public void setRepositoryURL(String repositoryURL) {
-        this.repositoryURL = repositoryURL;
-    }
-
-    public Properties getLabels() {
-        return labels;
-    }
-
-    public void setLabels(Properties labels) {
-        this.labels = labels;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#hashCode()
-     */
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#equals(java.lang.Object)
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (!(obj instanceof ArchetypeVersion)) {
-            return false;
-        }
-        ArchetypeVersion other = (ArchetypeVersion) obj;
-        if (id == null) {
-            if (other.id != null) {
-                return false;
-            }
-        } else if (!id.equals(other.id)) {
-            return false;
-        }
-        return true;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#toString()
-     */
-    @Override
-    public String toString() {
-        return "ArchetypeVersion [id=" + id + "]";
-    }
+    public Properties getLabels();
 
 }
