+++ src/main/java/org/jboss/jdf/stacks/model/Bom.java
@@ -14,120 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.jboss.jdf.stacks.model;
 
-public class Bom {
+public interface Bom {
 
-    private String id;
+    public String getId();
 
-    private String name;
+    public String getName();
 
-    private String description;
+    public String getDescription();
 
-    private String groupId = "org.jboss.bom";
+    public String getGroupId();
 
-    private String artifactId;
+    public String getArtifactId();
 
-    private String recommendedVersion;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        if (groupId != null) {
-            this.groupId = groupId;
-        }
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getRecommendedVersion() {
-        return recommendedVersion;
-    }
-
-    public void setRecommendedVersion(String recommendedVersion) {
-        this.recommendedVersion = recommendedVersion;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#hashCode()
-     */
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#equals(java.lang.Object)
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (!(obj instanceof Bom)) {
-            return false;
-        }
-        Bom other = (Bom) obj;
-        if (id == null) {
-            if (other.id != null) {
-                return false;
-            }
-        } else if (!id.equals(other.id)) {
-            return false;
-        }
-        return true;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#toString()
-     */
-    @Override
-    public String toString() {
-        return "Bom [id=" + id + "]";
-    }
+    public String getRecommendedVersion();
 
 }
