+++ src/main/java/org/jboss/jdf/stacks/model/Runtime.java
@@ -14,187 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.jboss.jdf.stacks.model;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
-public class Runtime {
+public interface Runtime {
 
-    private String id;
+    public String getId();
 
-    private String name;
+    public String getName();
 
-    private String groupId;
+    public String getGroupId();
 
-    private String artifactId;
+    public String getArtifactId();
 
-    private String version;
+    public String getVersion();
 
-    private String url;
+    public String getUrl();
 
-    private String downloadUrl;
+    public String getDownloadUrl();
 
-    private Properties labels = new Properties();
+    public Properties getLabels();
 
-    private List<BomVersion> boms = new ArrayList<BomVersion>();
+    public List<BomVersion> getBoms();
 
-    private BomVersion defaultBom;
+    public BomVersion getDefaultBom();
 
-    private List<ArchetypeVersion> archetypes = new ArrayList<ArchetypeVersion>();
+    public List<ArchetypeVersion> getArchetypes();
 
-    private ArchetypeVersion defaultArchetype;
+    public ArchetypeVersion getDefaultArchetype();
 
-    private String license;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getDownloadUrl() {
-        return downloadUrl;
-    }
-
-    public void setDownloadUrl(String downloadUrl) {
-        this.downloadUrl = downloadUrl;
-    }
-
-    public Properties getLabels() {
-        return labels;
-    }
-
-    public void setLabels(Properties labels) {
-        this.labels = labels;
-    }
-
-    public List<BomVersion> getBoms() {
-        return boms;
-    }
-
-    public void setBoms(List<BomVersion> boms) {
-        this.boms = boms;
-    }
-
-    public BomVersion getDefaultBom() {
-        return defaultBom;
-    }
-
-    public void setDefaultBom(BomVersion defaultBom) {
-        this.defaultBom = defaultBom;
-    }
-
-    public List<ArchetypeVersion> getArchetypes() {
-        return archetypes;
-    }
-
-    public void setArchetypes(List<ArchetypeVersion> archetypes) {
-        this.archetypes = archetypes;
-    }
-
-    public ArchetypeVersion getDefaultArchetype() {
-        return defaultArchetype;
-    }
-
-    public void setDefaultArchetype(ArchetypeVersion defaultArchetype) {
-        this.defaultArchetype = defaultArchetype;
-    }
-
-    public String getLicense() {
-        return license;
-    }
-
-    public void setLicense(String license) {
-        this.license = license;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#hashCode()
-     */
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#equals(java.lang.Object)
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (!(obj instanceof Runtime)) {
-            return false;
-        }
-        Runtime other = (Runtime) obj;
-        if (id == null) {
-            if (other.id != null) {
-                return false;
-            }
-        } else if (!id.equals(other.id)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "Runtime [id=" + id + "]";
-    }
+    public String getLicense();
 
 }
