+++ src/main/resources/stacks.yaml
@@ -1,22 +1,18 @@
-#JBoss, Home of Professional Open Source
-# Copyright 2011, Red Hat, Inc., and individual contributors
-# by the @authors tag. See the copyright.txt in the distribution for a
-# full listing of individual contributors.
+# JBoss, Home of Professional Open Source
+# Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
+# contributors by the @authors tag. See the copyright.txt in the
+# distribution for a full listing of individual contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.#All Rules
 
-# This is free software; you can redistribute it and/or modify it
-# under the terms of the GNU Lesser General Public License as
-# published by the Free Software Foundation; either version 2.1 of
-# the License, or (at your option) any later version.
-
-# This software is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-# Lesser General Public License for more details.
-
-# You should have received a copy of the GNU Lesser General Public
-# License along with this software; if not, write to the Free
-# Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-# 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 
 licenses:
  - &lgpl http://www.gnu.org/copyleft/lesser.txt
@@ -254,7 +250,7 @@ availableBoms:
    ###############################
  - &jboss-javaee-6_0-with-spring
    id: jboss-javaee-6_0-with-spring
-   name: JJava EE 6 with Spring
+   name: Java EE 6 with Spring
    description: Dependency Management for Java EE 6 Specification APIs with Spring
    groupId: org.jboss.bom
    artifactId: jboss-javaee-6.0-with-spring
